change semantics, improve endTimepoint computation
This commit changes the semantics by adding guards to
the semantics of ifElse and while statements.
It also removes timepoints denoting the end of
an if-else branch or the end of the while-body,
which are not needed anymore because of the change.
The changed semantics simplify the soundness proofs.

The changes in this commit change the reasoning of
Vampire. There should be no unrolling of the loop
anymore (this is good), but due to additional
guards we lose many applications of demodulation (this
is bad).
As far as I understand, we get back all these
simplications as soon as subsumption demodulation
is implemented.