add option -nat to output native sort Nat
This commit adds an option, which can be activated using -nat on.
If it is activated, Rapid doesn't encode Nat as datatype, but as a
natively known sort which is declared as
(declare-nat Nat zero s p Sub)

Note:
- this encoding is an extension of the smtlib-syntax.
- Vampire's master branch currently (as of January 2020) does
  not support this encoding (but a private branch does).

The advantage of using such an extension is that Vampire explicitly
knows about Nat. It therefore can add internal theory axioms for Nat
(which at least currently are handled differently than external theory
axioms), and use dedicated inference rules to reason about Nat.