add new example: harder version of increment_by_one
Since we now have a stronger version of the ValueEvolution trace lemma,
it is possible to prove many examples without the Injectivity trace
lemma, in particular those where it would not matter if we visit an
array position more than once (since we e.g. always set the same value
to the same position).
This new example shows that we still sometimes need the injectivity
lemma (it does matter how often we add 1 to a position).