add support for lemma-predicates
We want to guide Vampire's literal selection function, so that lemmas
can always be proven before they are used (which is not the case by
default if the lemma's proof needs a positive equality, since positive
equalities are selected last in Vampire (at least in all non-total
complete literal-selection functions)). We have extended Vampire to use
such a literal-selection function for all predicates declared using the
custom smtlib2-extension (declare-lemma-predicate ...).

This commit changes the output to use this extension for all definitions
introduced by Rapid (if the option '-lemmaPredicates' is set to 'on').