change relational-semantics: use separate semantics for each trace
This commit changes the semantics produced by Rapid in the setting of
multiple traces. Instead of producing semantics of the form
	forall trace. [P](trace)
we now produce semantics of the form
[P](trace_1) and ... and [P](trace_k),
where k is the number of traces.

While this change induces an increase of axioms for the semantics which
is linear in the number of traces, it seems to payoff. In particular, it
seems that (at least after implementing a similar change for trace
lemmas) the decreased amount of quantification enables more efficient
reasoning in Vampire. This could be the case because more
simplifications can be performed since we don't need to instantiate the
universally quantified trace-variables during saturation. Another
advantage is that clauses are more often splittable, which only helps if
one uses Avatar.
More abstractly, it seems to payoff to instantiate universally-quantified
variables in Vampire, if the only terms of that sort occuring in the
encoding are nullary function-symbols.