improve theory axioms added by Rapid
This commit improves the theory axioms, which Rapid uses to axiomatize
the less-relation for natural numbers, if no native theory of natural
numbers is used. In particular, this enables somewhat fair comparisions
with other provers, which support SMTLIB, but don't support our custom
extensions of Vampire.

The added axioms now match the axioms added internally by the custom
version of Vampire used for Rapid.