implement InlinedVariableValues data structure and use it in trace lemmas
We use the SemanticsInliner to inline as many terms denoting
variable-values as possible. We now need to ensure that the inlining
does not lead to incompletess for other formulas. There are two other
types of formulas,
(i) user-conjectures: any term occuring in a user conjecture is called
a persistent term and will not be inlined
(ii) trace lemmas: for these, we are currently incomplete.

This commit fixes the incompleteness in trace lemmas, by keeping record
of the inlined terms using the InlinedVariableValues data structure,
and uses this data structure while generating the trace lemmas.
As a result, the generated trace lemmas also perform the inlining, which
resolves the incompleteness.